move homepage pie chart to a separate file and convert from class to func comp
