+++ src/components/Homepage/Homepage.tsx
@@ -1,9 +1,10 @@
-import React, { Component } from "react";
+import React, { ReactElement } from "react";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import Button from "react-bootstrap/Button";
 import CanvasJSReact from "../../assets/canvasjs.react";
+import { options } from "./chartConfigs";
 import "./homepage.css";
 
 const CanvasJS = CanvasJSReact.CanvasJS;
@@ -11,45 +12,9 @@ const CanvasJSChart = CanvasJSReact.CanvasJSChart;
 
 interface Props {}
 
-interface State {}
-
-export default class Homepage extends Component<Props, State> {
-  state = {};
-
-  render() {
-    const options = {
-      exportEnabled: true,
-      animationEnabled: true,
-      title: {
-        text: "Budget - March 2020: €2,235.00",
-        fontFamily: "Merriweather"
-      },
-
-      data: [
-        {
-          type: "pie",
-          // showInLegend: "true",
-          legendText: "{label}",
-          indexLabelFontSize: 16,
-          indexLabel: "{label} - #percent%",
-          percentFormatString: "#0.##",
-          yValueFormatString: "€#,###.00",
-          indexLabelFontFamily: "Caladea",
-          toolTipContent: "{y} (#percent%)",
-          dataPoints: [
-            { y: 670.5, label: "Savings Target" },
-            { y: 340.84, label: "Transportation" },
-            { y: 657.09, label: "Rent" },
-            { y: 234.68, label: "Groceries" },
-            { y: 93.87, label: "Clothing" },
-            { y: 111.75, label: "Donations" },
-            { y: 46.94, label: "Misc." },
-            { y: 79.34, label: "Remaining Budget" }
-          ]
-        }
-      ]
-    };
-    return (
+export default function Homepage({}: Props): ReactElement {
+  return (
+    <div>
       <Container fluid>
         <Row className="title">
           <Col>
@@ -85,6 +50,6 @@ export default class Homepage extends Component<Props, State> {
           </Col>
         </Row>
       </Container>
-    );
-  }
+    </div>
+  );
 }
